feat: Enhance hero carousel and add preloader
Adjusted hero carousel for longer slide durations (8s) and increased height to 800px for better visual appeal. Implemented a Ken Burns effect on carousel images with a 12s transition.

Introduced a preloader component to the App, visible for 2.5 seconds on initial load to provide a smoother user experience. Also, enhanced the Impact page with a new scanline effect for a more dynamic visual.